+++ frontend/src/App/index.tsx
@@ -19,7 +19,7 @@ const ProjectDetailsPaymentForm = lazy(() => import("src/pages/ProjectDetails/Pa
 
 import { CustomUserRole, HasuraUserRole } from "src/types";
 import LoaderFallback from "src/components/Loader";
-import ScrollToTop from "src/components/ScrollToTop";
+import ScrollToTop from "src/App/ScrollToTop";
 import ErrorTrigger from "src/pages/ErrorTrigger";
 
 export enum RoutePaths {
